+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Convenience function `partial-recall-current-moment`.
+- Lighter now shows focus using `partial-recall-graph` if fleeting.
 
 ## [0.9.1]